# packages/contracts/openapi.yaml
openapi: 3.1.0
...
info:
  title: Core Accounting MVP API
  version: 1.2.0
  description: 'مواصفات REST API للنواة المحاسبية القابلة للتوسع.


    مصادر التصميم: ERD v1.2، قاموس البيانات v1.2، مصفوفة الصلاحيات وقواعد الترحيل v1.1، وحالات الاستخدام وتدفقات العمل v1.0، مع تعديلات ADR-002 v1.3.


    قرارات تنفيذية مثبتة في هذه النسخة:

    - الشركة الحالية تُثبت في الجلسة عبر /auth/context، ولا يُعتمد companyId الوارد من العميل في العمليات التشغيلية.

    - رقم المستند يُحجز عند إنشاء المسودة مع قبول فجوات الأرقام الموثقة.

    - منشئ القيد اليدوي لا يرحله افتراضياً، ويمكن جعل ذلك إعداداً للشركات الصغيرة.

    - واجهة التخصيصات مدعومة في API، ولو أُجلت واجهتها في الإصدار الأول.

    - مدة الاحتفاظ بمفتاح Idempotency هي 24 ساعة افتراضياً وقابلة للضبط.

    - أقصى نطاق تفاعلي للتقارير 366 يوماً.


    المبالغ وأسعار الصرف والمعرفات الكبيرة تُنقل كنصوص للحفاظ على الدقة. الحقول المشفرة وBlind Indexes لا تظهر في الاستجابات
    العامة.


    تغييرات v1.1:

    - اعتماد counter_account_id رسمياً في سندات القبض والصرف العامة.

    - إضافة GET /dashboard/summary ببيانات مشتقة وصلاحيات حقلية.

    - لا تُخزن مؤشرات لوحة المتابعة كمصدر حقيقة مستقل.'
  contact:
    name: Core Accounting MVP Technical Team
  license:
    name: Internal Use
    identifier: LicenseRef-Internal
servers:
- url: /api/v1
  description: مسار نسبي يعمل على نفس أصل الواجهة في جميع البيئات
...
security:
- SessionCookie: []
paths:
...
  /payments:
    get:
      tags:
      - Payments
      summary: عرض سندات الصرف
      operationId: listPayments
      responses:
        '200':
          description: قائمة النتائج
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                - meta
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Payment'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessRuleViolation'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      - name: status
        in: query
        schema:
          $ref: '#/components/schemas/AccountingDocumentStatus'
      - name: dateFrom
        in: query
        schema:
          type: string
          format: date
          example: '2026-07-30'
      - name: dateTo
        in: query
        schema:
          type: string
          format: date
          example: '2026-07-30'
      - $ref: '#/components/parameters/Search'
      x-permission: payments.view
      security:
      - SessionCookie: []
    post:
      tags:
      - Payments
      summary: إنشاء سند صرف مسودة
      operationId: createPayment
      responses:
        '201':
          description: نجاح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessRuleViolation'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
      description: يحجز رقم المستند عند إنشاء المسودة، ويثبت اسم وعنوان الطرف في حقول Snapshot.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentCreateRequest'
      x-permission: payments.create
      security:
      - SessionCookie: []
        CsrfToken: []
      parameters:
      - $ref: '#/components/parameters/RequestId'
  /payments/{paymentId}:
    get:
      tags:
      - Payments
      summary: عرض سند صرف
      operationId: getPayment
      responses:
        '200':
          description: نجاح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessRuleViolation'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - name: paymentId
        in: path
        required: true
        description: معرف السند
        schema:
          type: string
          pattern: ^[1-9][0-9]*$
          example: '1001'
          description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
      x-permission: payments.view
      security:
      - SessionCookie: []
    patch:
      tags:
      - Payments
      summary: تعديل سند صرف مسودة
      operationId: updatePayment
      responses:
        '200':
          description: نجاح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessRuleViolation'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - name: paymentId
        in: path
        required: true
        description: معرف السند
        schema:
          type: string
          pattern: ^[1-9][0-9]*$
          example: '1001'
          description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentUpdateRequest'
      x-permission: payments.update
      security:
      - SessionCookie: []
        CsrfToken: []
  /payments/{paymentId}/post:
    post:
      tags:
      - Payments
      summary: ترحيل السند
      operationId: postPayment
      responses:
        '200':
          description: نجاح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentCommandResult'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessRuleViolation'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - name: paymentId
        in: path
        required: true
        description: معرف السند
        schema:
          type: string
          pattern: ^[1-9][0-9]*$
          example: '1001'
          description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostDocumentRequest'
      x-permission: payments.post
      security:
      - SessionCookie: []
        CsrfToken: []
      x-idempotency-scope: user + company + operation + key
      x-idempotency-retention: 24h (configurable)
  /payments/{paymentId}/cancel:
    post:
      tags:
      - Payments
      summary: إلغاء السند المسودة
      operationId: cancelPayment
      responses:
        '200':
          description: نجاح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentCommandResult'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessRuleViolation'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - name: paymentId
        in: path
        required: true
        description: معرف السند
        schema:
          type: string
          pattern: ^[1-9][0-9]*$
          example: '1001'
          description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelDocumentRequest'
      x-permission: payments.cancel
      security:
      - SessionCookie: []
        CsrfToken: []
  /payments/{paymentId}/reverse:
    post:
      tags:
      - Payments
      summary: عكس السند المرحل
      operationId: reversePayment
      responses:
        '200':
          description: نجاح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentCommandResult'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessRuleViolation'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - name: paymentId
        in: path
        required: true
        description: معرف السند
        schema:
          type: string
          pattern: ^[1-9][0-9]*$
          example: '1001'
          description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReverseDocumentRequest'
      x-permission: payments.reverse
      security:
      - SessionCookie: []
        CsrfToken: []
      x-idempotency-scope: user + company + operation + key
      x-idempotency-retention: 24h (configurable)
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: sid
      description: معرف جلسة عشوائي؛ القيمة الخام لا تُخزن في قاعدة البيانات.
    CsrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: رمز Synchronizer مرتبط بالجلسة ومطلوب لكل طلب يغير الحالة.
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    Search:
      name: search
      in: query
      schema:
        type: string
        maxLength: 200
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: UUID أو قيمة عشوائية فريدة للعملية. تُخزن النتيجة 24 ساعة افتراضياً.
      schema:
        type: string
        minLength: 16
        maxLength: 100
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      schema:
        type: string
        format: uuid
  responses:
    ValidationError:
      description: خطأ في التنسيق أو الحقول
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthenticated:
      description: الجلسة مفقودة أو منتهية أو ملغاة
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: لا توجد صلاحية أو سياق الشركة غير مصرح به
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: الكيان غير موجود داخل الشركة الحالية؛ لا يكشف وجوده في شركة أخرى
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: تعارض إصدار أو تزامن أو تكرار
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    BusinessRuleViolation:
      description: انتهاك قاعدة أعمال، مثل فترة مغلقة أو قيد غير متوازن
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    RateLimited:
      description: تم تجاوز معدل الطلبات
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    InternalError:
      description: خطأ داخلي؛ يعاد requestId فقط دون تفاصيل حساسة
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    PaginationMeta:
      type: object
      required:
      - page
      - pageSize
      - totalItems
      - totalPages
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 100
        totalItems:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    Problem:
      type: object
      required:
      - code
      - messageAr
      - requestId
      properties:
        code:
          type: string
          enum:
          - VALIDATION_ERROR
          - UNAUTHENTICATED
          - FORBIDDEN
          - NOT_FOUND
          - CONFLICT
          - BUSINESS_RULE_VIOLATION
          - RATE_LIMITED
          - INTERNAL_ERROR
        messageAr:
          type: string
        requestId:
          type: string
          format: uuid
        fieldErrors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        details:
          type: object
          additionalProperties: true
          description: تفاصيل آمنة فقط؛ لا تتضمن SQL أو Stack Trace أو قيماً حساسة.
...
    AccountingDocumentStatus:
      type: string
      enum:
      - DRAFT
      - POSTED
      - CANCELLED
      - REVERSED
    AccountingDocumentType:
      type: string
      enum:
      - MANUAL_JOURNAL
      - RECEIPT
      - PAYMENT
      - SALES_INVOICE
      - SALES_CREDIT_NOTE
      - PURCHASE_INVOICE
      - PURCHASE_DEBIT_NOTE
      - PERIOD_CLOSE
    AccountingDocumentHeader:
      type: object
      required:
      - id
      - documentType
      - documentNumber
      - documentDate
      - description
      - status
      - fiscalPeriodId
      - version
      - createdBy
      - createdAt
      properties:
        id:
          type: string
          pattern: ^[1-9][0-9]*$
          example: '1001'
          description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
        documentType:
          $ref: '#/components/schemas/AccountingDocumentType'
        documentNumber:
          type: string
          maxLength: 60
        documentDate:
          type: string
          format: date
          example: '2026-07-30'
        description:
          type: string
          maxLength: 500
        status:
          $ref: '#/components/schemas/AccountingDocumentStatus'
        fiscalPeriodId:
          type: string
          pattern: ^[1-9][0-9]*$
          example: '1001'
          description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
        version:
          type: integer
          minimum: 0
          example: 0
        createdBy:
          type: string
          pattern: ^[1-9][0-9]*$
          example: '1001'
          description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
        createdAt:
          type: string
          format: date-time
          example: '2026-07-30T06:00:00.000Z'
        postedBy:
          oneOf:
          - type: string
            pattern: ^[1-9][0-9]*$
            example: '1001'
            description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
          - type: 'null'
        postedAt:
          oneOf:
          - type: string
            format: date-time
            example: '2026-07-30T06:00:00.000Z'
          - type: 'null'
        reversedByDocumentId:
          oneOf:
          - type: string
            pattern: ^[1-9][0-9]*$
            example: '1001'
            description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
          - type: 'null'
...
    PostDocumentRequest:
      type: object
      additionalProperties: false
      required:
      - version
      properties:
        version:
          type: integer
          minimum: 0
          example: 0
    CancelDocumentRequest:
      type: object
      additionalProperties: false
      required:
      - version
      - reason
      properties:
        version:
          type: integer
          minimum: 0
          example: 0
        reason:
          type: string
          minLength: 3
          maxLength: 500
    ReverseDocumentRequest:
      type: object
      additionalProperties: false
      required:
      - version
      - reversalDate
      - reason
      properties:
        version:
          type: integer
          minimum: 0
          example: 0
        reversalDate:
          type: string
          format: date
          example: '2026-07-30'
        reason:
          type: string
          minLength: 3
          maxLength: 500
    DocumentCommandResult:
      type: object
      required:
      - document
      - requestId
      properties:
        document:
          $ref: '#/components/schemas/AccountingDocumentHeader'
        generatedJournalEntryIds:
          type: array
          items:
            type: string
            pattern: ^[1-9][0-9]*$
            example: '1001'
            description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
        requestId:
          type: string
          format: uuid
        idempotentReplay:
          type: boolean
          default: false
    AllocationInput:
      type: object
      additionalProperties: false
      required:
      - targetJournalLineId
      - allocatedAmount
      properties:
        targetJournalLineId:
          type: string
          pattern: ^[1-9][0-9]*$
          example: '1001'
          description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
        allocatedAmount:
          type: string
          pattern: ^(0|[1-9][0-9]{0,14})\.[0-9]{4}$
          example: '1250.0000'
          description: قيمة مالية عشرية DECIMAL(19,4) كنص.
    Allocation:
      allOf:
      - $ref: '#/components/schemas/AllocationInput'
      - type: object
        required:
        - id
        properties:
          id:
            type: string
            pattern: ^[1-9][0-9]*$
            example: '1001'
            description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
...
    PaymentCreateRequest:
      type: object
      additionalProperties: false
      required:
      - fiscalPeriodId
      - documentDate
      - description
      - cashBankAccountId
      - paymentMethodId
      - currencyId
      - exchangeRate
      - amount
      - counterpartyName
      properties:
        fiscalPeriodId:
          type: string
          pattern: ^[1-9][0-9]*$
          example: '1001'
          description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
        documentDate:
          type: string
          format: date
          example: '2026-07-30'
        description:
          type: string
          minLength: 1
          maxLength: 500
        supplierId:
          oneOf:
          - type: string
            pattern: ^[1-9][0-9]*$
            example: '1001'
            description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
          - type: 'null'
          description: في PATCH يجوز إغفاله. بعد دمج التعديل مع السجل يجب أن تكون إحدى القيمتين supplierId أو counterAccountId غير فارغة حصراً.
        counterAccountId:
          oneOf:
          - type: string
            pattern: ^[1-9][0-9]*$
            example: '1001'
            description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
          - type: 'null'
          description: إلزامي إذا لم يوجد supplierId.
        cashBankAccountId:
          type: string
          pattern: ^[1-9][0-9]*$
          example: '1001'
          description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
        paymentMethodId:
          type: string
          pattern: ^[1-9][0-9]*$
          example: '1001'
          description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
        currencyId:
          type: string
          pattern: ^[1-9][0-9]*$
          example: '1001'
          description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
        exchangeRate:
          type: string
          pattern: ^[0-9]{1,11}\.[0-9]{8}$
          example: '1.00000000'
          description: سعر صرف DECIMAL(19,8) كنص ويجب أن يكون أكبر من صفر.
        amount:
          type: string
          pattern: ^(0|[1-9][0-9]{0,14})\.[0-9]{4}$
          example: '1250.0000'
          description: قيمة مالية عشرية DECIMAL(19,4) كنص.
        referenceNumber:
          type:
          - string
          - 'null'
          maxLength: 100
        counterpartyName:
          type: string
          minLength: 1
          maxLength: 200
        counterpartyTaxNumber:
          type:
          - string
          - 'null'
          writeOnly: true
        counterpartyAddress:
          type:
          - string
          - 'null'
          maxLength: 500
        notes:
          type:
          - string
          - 'null'
          maxLength: 1000
        allocations:
          type: array
          items:
            $ref: '#/components/schemas/AllocationInput'
      description: يجب إرسال أحد supplierId أو counterAccountId فقط.
      oneOf:
      - required:
        - supplierId
        not:
          type: object
          properties:
            counterAccountId: {}
          required:
          - counterAccountId
      - required:
        - counterAccountId
        not:
          type: object
          properties:
            supplierId: {}
          required:
          - supplierId
    PaymentUpdateRequest:
      type: object
      additionalProperties: false
      required:
      - version
      properties:
        fiscalPeriodId:
          type: string
          pattern: ^[1-9][0-9]*$
          example: '1001'
          description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
        documentDate:
          type: string
          format: date
          example: '2026-07-30'
        description:
          type: string
          minLength: 1
          maxLength: 500
        supplierId:
          oneOf:
          - type: string
            pattern: ^[1-9][0-9]*$
            example: '1001'
            description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
          - type: 'null'
        counterAccountId:
          oneOf:
          - type: string
            pattern: ^[1-9][0-9]*$
            example: '1001'
            description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
          - type: 'null'
          description: إلزامي إذا لم يوجد supplierId.
        cashBankAccountId:
          type: string
          pattern: ^[1-9][0-9]*$
          example: '1001'
          description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
        paymentMethodId:
          type: string
          pattern: ^[1-9][0-9]*$
          example: '1001'
          description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
        currencyId:
          type: string
          pattern: ^[1-9][0-9]*$
          example: '1001'
          description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
        exchangeRate:
          type: string
          pattern: ^[0-9]{1,11}\.[0-9]{8}$
          example: '1.00000000'
          description: سعر صرف DECIMAL(19,8) كنص ويجب أن يكون أكبر من صفر.
        amount:
          type: string
          pattern: ^(0|[1-9][0-9]{0,14})\.[0-9]{4}$
          example: '1250.0000'
          description: قيمة مالية عشرية DECIMAL(19,4) كنص.
        referenceNumber:
          type:
          - string
          - 'null'
          maxLength: 100
        counterpartyName:
          type: string
          minLength: 1
          maxLength: 200
        counterpartyTaxNumber:
          type:
          - string
          - 'null'
          writeOnly: true
        counterpartyAddress:
          type:
          - string
          - 'null'
          maxLength: 500
        notes:
          type:
          - string
          - 'null'
          maxLength: 1000
        allocations:
          type: array
          items:
            $ref: '#/components/schemas/AllocationInput'
        version:
          type: integer
          minimum: 0
          example: 0
      description: يجوز إغفال حقلي الطرف في التحديث الجزئي؛ وعند إرسالهما يجب أن تنتج القيم المدمجة طرفًا واحدًا فقط من supplierId أو counterAccountId.
    Payment:
      type: object
      required:
      - id
      - document
      - cashBankAccountId
      - paymentMethodId
      - currencyId
      - exchangeRate
      - amount
      - baseAmount
      - counterpartyNameSnapshot
      - allocations
      properties:
        id:
          type: string
          pattern: ^[1-9][0-9]*$
          example: '1001'
          description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
        document:
          $ref: '#/components/schemas/AccountingDocumentHeader'
        supplierId:
          oneOf:
          - type: string
            pattern: ^[1-9][0-9]*$
            example: '1001'
            description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
          - type: 'null'
        counterAccountId:
          oneOf:
          - type: string
            pattern: ^[1-9][0-9]*$
            example: '1001'
            description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
          - type: 'null'
        cashBankAccountId:
          type: string
          pattern: ^[1-9][0-9]*$
          example: '1001'
          description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
        paymentMethodId:
          type: string
          pattern: ^[1-9][0-9]*$
          example: '1001'
          description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
        currencyId:
          type: string
          pattern: ^[1-9][0-9]*$
          example: '1001'
          description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
        exchangeRate:
          type: string
          pattern: ^[0-9]{1,11}\.[0-9]{8}$
          example: '1.00000000'
          description: سعر صرف DECIMAL(19,8) كنص ويجب أن يكون أكبر من صفر.
        amount:
          type: string
          pattern: ^(0|[1-9][0-9]{0,14})\.[0-9]{4}$
          example: '1250.0000'
          description: قيمة مالية عشرية DECIMAL(19,4) كنص.
        baseAmount:
          type: string
          pattern: ^(0|[1-9][0-9]{0,14})\.[0-9]{4}$
          example: '1250.0000'
          description: قيمة مالية عشرية DECIMAL(19,4) كنص.
        referenceNumber:
          type:
          - string
          - 'null'
        counterpartyNameSnapshot:
          type: string
        counterpartyTaxMasked:
          type:
          - string
          - 'null'
        counterpartyAddressSnapshot:
          type:
          - string
          - 'null'
        notes:
          type:
          - string
          - 'null'
        allocations:
          type: array
          items:
            $ref: '#/components/schemas/Allocation'
      description: يطبق XOR بين supplierId وcounterAccountId.
...